+++ CHANGELOG.md
@@ -9,6 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Take Save out of "more" menu
 
+## [1.113] - 2022-12-25
+### Added
+- Email verification support. If verification is supported by server, the new
+  field (Verification code) is appear in Sign Up and Profile forms. When user
+  sets up or updates email address, they should receive verification code on it
+  and enter that code to the form.
+- NodeJS 18.x is supported now
+
+### Changed
+- "Everything", "Search" and "Best of" pages are not available for
+  non-authorized users.
+- Updated texts about donations to FreeFeed MTÜ
+
 ## [1.112] - 2022-11-25
 ### Added
 - Group administrators can now block users in their managed groups. A blocked